+++ widgets/views/snippet_markdown.php
@@ -1,4 +1,6 @@
 <?php
+
+use humhub\modules\content\widgets\richtext\RichText;
 use humhub\modules\custom_pages\widgets\SnippetContent;
 
 /* @var $model \humhub\modules\custom_pages\models\CustomContentContainer */
@@ -8,8 +10,8 @@ $navigation = (!$canEdit) ? [] : [
     '<a href="'.$model->getEditUrl().'"><i class="fa fa-pencil"></i>' . Yii::t('CustomPagesModule.base', 'Edit') . '</a>'
 ];
 ?>
-<?= \humhub\modules\custom_pages\widgets\SnippetContent::widget([
+<?= SnippetContent::widget([
     'model' => $model,
-    'content' => humhub\widgets\MarkdownView::widget(['markdown' => $model->getPageContent()]),
+    'content' => RichText::output( $model->getPageContent()),
     'navigation' => $navigation
 ]); ?>